Update the events and members cid if club cid changes. Minor other changes
