+++ mutations_members.py
@@ -4,6 +4,7 @@ from fastapi.encoders import jsonable_encoder
 from datetime import datetime
 
 from db import membersdb
+from utils import unique_roles_id, non_deleted_members
 
 # import all models and types
 from otypes import Info
@@ -13,74 +14,6 @@ from otypes import FullMemberInput, SimpleMemberInput, MemberType
 """
 MEMBER MUTATIONS
 """
-
-
-def non_deleted_members(member_input) -> MemberType:
-    """
-    Function to return non-deleted members for a particular cid, uid
-    Only to be used in admin functions, as it returns both approved/non-approved members.
-    """
-    updated_sample = membersdb.find_one(
-        {
-            "$and": [
-                {"cid": member_input["cid"]},
-                {"uid": member_input["uid"]},
-            ]
-        },
-        {"_id": 0},
-    )
-    if updated_sample is None:
-        raise Exception("No such Record")
-
-    roles = []
-    for i in updated_sample["roles"]:
-        if i["deleted"] is True:
-            continue
-        roles.append(i)
-    updated_sample["roles"] = roles
-
-    return MemberType.from_pydantic(Member.parse_obj(updated_sample))
-
-
-def unique_roles_id(uid, cid):
-    """
-    Function to give unique ids for each of the role in roles list
-    """
-    pipeline = [
-        {
-            "$set": {
-                "roles": {
-                    "$map": {
-                        "input": {"$range": [0, {"$size": "$roles"}]},
-                        "in": {
-                            "$mergeObjects": [
-                                {"$arrayElemAt": ["$roles", "$$this"]},
-                                {
-                                    "rid": {
-                                        "$toString": {
-                                            "$add": [
-                                                {"$toLong": datetime.now()},
-                                                "$$this",
-                                            ]
-                                        }
-                                    }
-                                },
-                            ]
-                        },
-                    }
-                }
-            }
-        }
-    ]
-    membersdb.update_one(
-        {
-            "$and": [
-                {"cid": cid},
-                {"uid": uid},
-            ]
-        },
-        pipeline,
-    )
 
 
 @strawberry.mutation
